refactor(phase2): 引数定義と設定ヘルパーを library/args.py に分離
train_util.py の add_*_arguments, verify_*, get_sanitized_config_or_none,
read_config_from_file, resume_from_local_or_hf_if_specified の計 12 関数を
library/args.py に移動。train_util.py は re-export shim として維持。
enable_high_vram は accelerator 系（PR-3 で accelerator_setup.py 行き）の
ため train_util.py に残し、verify_training_args からは関数内 import で参照。

train_util.py: 3,774 → 2,600 行 (1,174 行削減)
library/args.py: 1,221 行（新規）
